pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thenvoi-sdk"
version = "0.1.2"
description = "A Python SDK for Thenvoi API"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "thenvoi" }]
license = { text = "MIT" }
keywords = ["sdk", "thenvoi", "api", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "thenvoi-client-rest==0.0.4",
    "phoenix-channels-python-client>=0.1.4",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0",
    "claude-agent-sdk>=0.1.16",
]

[project.scripts]
thenvoi-quality-gates = "thenvoi.testing.quality_gates:main"

[project.optional-dependencies]
codex = [
    "websockets>=13.0",
]
pydantic-ai = [
    "pydantic-ai-slim>=1.0.0",
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.75.0",
]
langgraph = [
    "langchain>=1.0.0",
    "langchain-core>=1.2.5",
    "langgraph>=1.0.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.4.0",  # 0.4.0+ required for langchain 1.0 compatibility
    "langchain-text-splitters>=0.3.0",
    "openai>=1.0.0",
    "beautifulsoup4>=4.12.0",
]
claude_sdk = [
    "claude-agent-sdk>=0.1.0",  # Requires Node.js 20+ and @anthropic-ai/claude-code CLI
]
parlant = [
    "parlant>=3.0.0",
    "openai>=1.0.0",
]
crewai = [
    # Upper bound due to reliance on kickoff_async message list behavior.
    # TODO: Test against CrewAI 1.0 when released and update upper bound if compatible.
    "crewai>=0.80.0,<2.0.0",
    "openai>=1.0.0",
    "nest-asyncio>=1.6.0",  # Required for sync-to-async bridging in CrewAI tools
]
a2a = [
    "a2a-sdk>=0.3.22",  # Official A2A Python SDK
]
a2a_gateway = [
    "a2a-sdk>=0.3.22",  # Official A2A Python SDK
    "starlette>=0.40.0",  # HTTP server framework
    "uvicorn>=0.32.0",  # ASGI server
]
a2a_gateway_demo = [
    "a2a-sdk>=0.3.22",      # A2A client for calling gateway peers
    "starlette>=0.40.0",    # HTTP server
    "uvicorn>=0.32.0",      # ASGI server
    "langgraph>=1.0.0",     # LangGraph for orchestrator agent
    "langchain-openai>=0.3.0",  # OpenAI model for LLM
    "click>=8.0.0",         # CLI for demo orchestrator
]
bridge = [
    "aiohttp>=3.9,<4",      # Health check HTTP server
]

dev = [
    "thenvoi-testing-python @ git+https://github.com/thenvoi/thenvoi-testing-python.git@thenvoi-testing-python-v0.1.1",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",  # Already in main deps, for mocking
    # Include pydantic-ai for testing
    "pydantic-ai-slim>=1.0.0",
    # Include anthropic for testing
    "anthropic>=0.75.0",
    # Include langgraph for testing
    "langchain>=1.0.0",
    "langchain-core>=1.2.5",
    "langgraph>=1.0.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.4.0",
    "langchain-text-splitters>=0.3.0",
    "openai>=1.0.0",
    "beautifulsoup4>=4.12.0",
    # Include parlant for testing
    "parlant>=3.0.0",
    # Include a2a-sdk for testing
    "a2a-sdk>=0.3.22",
    # Include a2a_gateway deps for testing
    "starlette>=0.40.0",
    "uvicorn>=0.32.0",
    # Include a2a_gateway_demo deps for testing
    "click>=8.0.0",
    # Include crewai for testing
    "crewai>=0.80.0",
    "nest-asyncio>=1.6.0",  # Required for CrewAI adapter sync-to-async bridging
    # Include bridge deps for testing
    "aiohttp>=3.9,<4",
    # Development tools
    "pre-commit>=3.0.0",
    "ruff>=0.8.0",
    "pyrefly>=0.18.0",
    "commitizen>=3.13.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["thenvoi", "thenvoi.*"]
namespaces = false

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v -m 'not integration'"
markers = [
    "integration: marks tests that exercise real external systems or end-to-end flows",
    "requires_api: integration tests requiring THENVOI_API_KEY",
    "requires_multi_agent: integration tests requiring THENVOI_API_KEY and THENVOI_API_KEY_2",
    "requires_user_api: integration tests requiring THENVOI_API_KEY_USER",
    "contract_gate: repository quality-gate contract checks required in CI",
]

[tool.pyrefly]
project_includes = [
    "src/**/*.py",              # Only check source files
]
project_excludes = [
    "tests/**",                 # Tests use mocks with duck typing
    "examples/**",              # Examples use local imports
    "thenvoi-bridge/**",        # Standalone bridge app, not in src/ import root
]
# Disable the default **/.[!/.]*/** exclude so pyrefly works inside
# .claude/worktrees/ (git worktree paths contain a hidden directory).
disable-project-excludes-heuristics = true

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.1"
version_files = [
    "pyproject.toml:version",
    "src/thenvoi/__init__.py:__version__",
]
# Tag format aligned with release-please (package-name-v$version)
tag_format = "thenvoi-sdk-python-v$version"

[tool.ruff]
